Don't start X services with systemd --user.
It turns out the there was a reason systemd --user wasn't all that well
documented. They hadn't actually got it to a start where it was doing
the right thing.

As version 206 it is doesn't work the why I was using it any more.
Instead of being launch on demand by the user it's now launched on the
first log in of a user. It still seems to work for MPD. It would
probably be OK for user daemons and stuff. Maybe I'll use it to manage
git annex and do some cron type stuff one day.